examples/openai_agent/requirements.txt
bunkervm[openai-agents]
python-dotenv
